# Add Shortest Path Faster Algorithm for calculating single source shortest path 