--- a/src/shortestpaths/spfa.jl
+++ b/src/shortestpaths/spfa.jl
@@ -16,6 +16,29 @@ Compute shortest paths between a source `s` and all
 other nodes in graph `g` using the [Shortest Path Faster Algorithm]
 (https://en.wikipedia.org/wiki/Shortest_Path_Faster_Algorithm).
 
+# Examples
+
+```jldoctest
+julia> g = CompleteGraph(3);
+
+julia> d = [1 -3 1; -3 1 1; 1 1 1];
+
+julia> spfa_shortest_paths(g, 1, d)
+
+ERROR: LightGraphs.NegativeCycleError()
+
+julia> g = CompleteGraph(4);
+
+julia> d = [1 1 -1 1; 1 1 -1 1; 1 1 1 1; 1 1 1 1];
+
+julia> spfa_shortest_paths(gx, 1, d)
+
+4-element Array{Int64,1}:
+  0
+  0
+ -1
+  0
+```
 
 """
 
@@ -74,15 +97,20 @@ Function which returns true if there is any negative weight cycle in the graph.
 
 ```jldoctest
 julia> g = CompleteGraph(3);
+
 julia> d = [1 -3 1; -3 1 1; 1 1 1];
+
 julia> has_negative_edge_cycle_spfa(g, d)
 true
 
 julia> g = CompleteGraph(4);
+
 julia> d = [1 1 -1 1; 1 1 -1 1; 1 1 1 1; 1 1 1 1];
+
 julia> has_negative_edge_cycle_spfa(g, d);
 false
 ```
+
 """
 function has_negative_edge_cycle_spfa(
     g::AbstractGraph{U},