--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -44,6 +44,7 @@ tests = [
     "shortestpaths/johnson",
     "shortestpaths/floyd-warshall",
     "shortestpaths/yen",
+    "shortestpaths/spfa",
     "traversals/bfs",
     "traversals/bipartition",
     "traversals/greedy_color",