--- a/test/shortestpaths/spfa.jl
+++ b/test/shortestpaths/spfa.jl
@@ -75,8 +75,8 @@
     @testset "Random Graphs" begin
         @testset "Simple graphs" begin
             for i = 1:5
-                nvg = Int(floor(250*rand()))
-                neg = Int(floor((nvg*(nvg-1)/2)*rand()))
+                nvg = Int(ceil(250*rand()))
+                neg = Int(ceil((nvg*(nvg-1)/2)*rand()))
                 seed = Int(floor(100*rand()))
                 g = SimpleGraph(nvg, neg; seed = seed)
                 z = spfa_shortest_paths(g, 1)