--- a/test/shortestpaths/spfa.jl
+++ b/test/shortestpaths/spfa.jl
@@ -76,7 +76,7 @@
         @testset "Simple graphs" begin
             for i = 1:5
                 nvg = Int(ceil(250*rand()))
-                neg = Int(ceil((nvg*(nvg-1)/2)*rand()))
+                neg = Int(floor((nvg*(nvg-1)/2)*rand()))
                 seed = Int(floor(100*rand()))
                 g = SimpleGraph(nvg, neg; seed = seed)
                 z = spfa_shortest_paths(g, 1)
@@ -87,8 +87,8 @@
 
         @testset "Simple DiGraphs" begin
             for i = 1:5
-                nvg = 10
-                neg = 45
+                nvg = Int(ceil(250*rand()))
+                neg = Int(floor((nvg*(nvg-1)/2)*rand()))
                 seed = Int(floor(100*rand()))
                 g = SimpleDiGraph(nvg, neg; seed = seed)
                 z = spfa_shortest_paths(g, 1)