--- a/test/shortestpaths/spfa.jl
+++ b/test/shortestpaths/spfa.jl
@@ -153,24 +153,26 @@
         end
     end
 
-    # Negative Cycle
-    gx = CompleteGraph(3)
-    for g in testgraphs(gx)
-        d = [1 -3 1; -3 1 1; 1 1 1]
-        @test_throws LightGraphs.NegativeCycleError spfa_shortest_paths(g, 1, d)
-        @test has_negative_edge_cycle_spfa(g, d)
+    @testset "Negative Cycle" begin
+        # Negative cycle 1 
+        gx = CompleteGraph(3)
+        for g in testgraphs(gx)
+            d = [1 -3 1; -3 1 1; 1 1 1]
+            @test_throws LightGraphs.NegativeCycleError spfa_shortest_paths(g, 1, d)
+            @test has_negative_edge_cycle_spfa(g, d)
 
-        d = [1 -1 1; -1 1 1; 1 1 1]
-        @test_throws LightGraphs.NegativeCycleError spfa_shortest_paths(g, 1, d)
-        @test has_negative_edge_cycle_spfa(g, d)
-    end
+            d = [1 -1 1; -1 1 1; 1 1 1]
+            @test_throws LightGraphs.NegativeCycleError spfa_shortest_paths(g, 1, d)
+            @test has_negative_edge_cycle_spfa(g, d)
+        end
 
-    # Negative cycle of length 3 in graph of diameter 4
-    gx = CompleteGraph(4)
-    d = [1 -1 1 1; 1 1 1 -1; 1 1 1 1; 1 1 1 1]
-    for g in testgraphs(gx)
-        @test_throws LightGraphs.NegativeCycleError spfa_shortest_paths(g, 1, d)
-        @test has_negative_edge_cycle_spfa(g, d)
-    end
+        # Negative cycle of length 3 in graph of diameter 4
+        gx = CompleteGraph(4)
+        d = [1 -1 1 1; 1 1 1 -1; 1 1 1 1; 1 1 1 1]
+        for g in testgraphs(gx)
+            @test_throws LightGraphs.NegativeCycleError spfa_shortest_paths(g, 1, d)
+            @test has_negative_edge_cycle_spfa(g, d)
+        end
+     end
 
 end