--- a/test/shortestpaths/spfa.jl
+++ b/test/shortestpaths/spfa.jl
@@ -153,8 +153,31 @@
         end
     end
 
+    @testset "Normal graph" begin
+        g4 = PathDiGraph(5)
+
+        d1 = float([0 1 2 3 4; 5 0 6 7 8; 9 10 0 11 12; 13 14 15 0 16; 17 18 19 20 0])
+        d2 = sparse(float([0 1 2 3 4; 5 0 6 7 8; 9 10 0 11 12; 13 14 15 0 16; 17 18 19 20 0]))
+        for g in testdigraphs(g4)
+            y = @inferred(spfa_shortest_paths(g, 2, d1))
+            z = @inferred(spfa_shortest_paths(g, 2, d2))
+            @test y == z == [Inf, 0, 6, 17, 33]
+            @test @inferred(!has_negative_edge_cycle_spfa(g))
+            @test @inferred(!has_negative_edge_cycle_spfa(g, d1))
+
+
+            y = @inferred(spfa_shortest_paths(g, 2, d1))
+            z = @inferred(spfa_shortest_paths(g, 2, d2))
+            @test y == z == [Inf, 0, 6, 17, 33]
+            @test @inferred(!has_negative_edge_cycle_spfa(g))
+            z = @inferred(spfa_shortest_paths(g, 2))
+            @test z == [typemax(Int), 0, 1, 2, 3]
+        end
+    end
+
+
     @testset "Negative Cycle" begin
-        # Negative cycle 1 
+        # Negative Cycle 1
         gx = CompleteGraph(3)
         for g in testgraphs(gx)
             d = [1 -3 1; -3 1 1; 1 1 1]
@@ -173,6 +196,6 @@
             @test_throws LightGraphs.NegativeCycleError spfa_shortest_paths(g, 1, d)
             @test has_negative_edge_cycle_spfa(g, d)
         end
-     end
+    end
 
 end