SPFA is an improvement of the Bellman-Ford algorithm which takes advantage of the fact that not all attempts at relaxation will work. The main idea is to create a queue containing only the vertices that were relaxed but that still could not relax their neighbors. And whenever you can relax some neighbor, you should put the neighbor in the queue. This algorithm work for negative edge weights and can also be used to detect negative cycles as the Bellman-Ford.

The worst case of this algorithm is equal to the `O(EV)` of the Bellman-Ford, but in practice it works much faster and it is found that it works even in `O(E)` on average. 

The only problem is that we can't find the parent vector through this algortihm. We can identify negative
weight cycles through this algortihm easily. `has_negative_cycle_spfa()` should be the goto function to check negative cycle in the graph because SPFA is faster and the way I have written it, it throws a `NegativeCycleError()` as soon as the negative cycle is detected. 
Contrary to this, the current implementation of Bellman Shortest path algorithm to detect negative cycle throws `NegativeCycleError()` after the `O(VE)` loop has completely finished. So this is slower.

Some benchmarks : 
```
julia> ee = SimpleGraph{Int32}(loadsnap(:email_enron_full))
{36692, 183831} undirected simple Int32 graph

julia> @benchmark spfa_shortest_paths(ee,1)
BenchmarkTools.Trial: 
  memory estimate:  722.84 KiB
  allocs estimate:  21
  --------------
  minimum time:     2.662 ms (0.00% GC)
  median time:      2.910 ms (0.00% GC)
  mean time:        2.988 ms (0.00% GC)
  maximum time:     5.158 ms (0.00% GC)
  --------------
  samples:          1643
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(ee,1)
BenchmarkTools.Trial: 
  memory estimate:  1.31 MiB
  allocs estimate:  142
  --------------
  minimum time:     5.218 ms (0.00% GC)
  median time:      5.552 ms (0.00% GC)
  mean time:        6.020 ms (1.62% GC)
  maximum time:     56.360 ms (90.36% GC)
  --------------
  samples:          829
  evals/sample:     1


julia> egu = SimpleGraph{Int32}(loadsnap(:ego_twitter_u))
{81306, 1342310} undirected simple Int32 graph

julia> @benchmark spfa_shortest_paths(egu,1)
BenchmarkTools.Trial: 
  memory estimate:  2.01 MiB
  allocs estimate:  23
  --------------
  minimum time:     14.253 ms (0.00% GC)
  median time:      14.670 ms (0.00% GC)
  mean time:        15.058 ms (0.00% GC)
  maximum time:     25.921 ms (0.00% GC)
  --------------
  samples:          328
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(egu,1)
BenchmarkTools.Trial: 
  memory estimate:  3.87 MiB
  allocs estimate:  153
  --------------
  minimum time:     23.616 ms (0.00% GC)
  median time:      25.901 ms (0.00% GC)
  mean time:        27.845 ms (1.43% GC)
  maximum time:     73.616 ms (67.55% GC)
  --------------
  samples:          180
  evals/sample:     1


julia> ss = SimpleDiGraph{Int32}(loadsnap(:soc_slashdot0902_d));

julia> @benchmark spfa_shortest_paths(ss,1)
BenchmarkTools.Trial: 
  memory estimate:  2.02 MiB
  allocs estimate:  23
  --------------
  minimum time:     7.871 ms (0.00% GC)
  median time:      8.119 ms (0.00% GC)
  mean time:        8.192 ms (0.00% GC)
  maximum time:     10.808 ms (0.00% GC)
  --------------
  samples:          601
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(ss,1)
BenchmarkTools.Trial: 
  memory estimate:  3.80 MiB
  allocs estimate:  158
  --------------
  minimum time:     15.901 ms (0.00% GC)
  median time:      16.422 ms (0.00% GC)
  mean time:        16.913 ms (1.53% GC)
  maximum time:     66.154 ms (74.78% GC)
  --------------
  samples:          296
  evals/sample:     1
```