@sbromberger 2019-04-09
Here are the two lines that need explicit test coverage: https://codecov.io/gh/JuliaGraphs/LightGraphs.jl/compare/d08ab6bb29ac41e0abbda7646f0332fa9c9aa6c3...e015134b1628fe2ae384b3783009e06d2abfd127/diff